Create ProfileLoaderAnimation component wo we can use it later
